Fix abigen with int expression template arguments
In my contracts, I use several template arguments which are types
which resolve to integral values (such as an eosio::name) but are
not trivially integer values. This works great in my contrats, but
causes abigen to fall on its face.

Fix abigen in cases where a template argument is an expression that
resolves to an integer, but isn't an integer literal.